Clean up and update color
